Struct sudachi::analysis::created::CreatedWords

pub struct CreatedWords(/* private fields */);

Bitset which represents that a word of a specified length was created. Lattice construction fills this bitmap and passes it to the OOV providers. It allows OOV providers to check if a word of a specific length was created very cheaply.

Unfortunately, if a word is more than MAX_VALUE characters, handlers need to do usual linear-time check.

Implementations

impl CreatedWords

pub const MAX_VALUE: u64 = 64u64

Maximum supported length of the word

pub fn empty() -> CreatedWords

pub fn single<Pos: Into<i64>>(length: Pos) -> CreatedWords

pub fn add_word<P: Into<i64>>(&self, length: P) -> CreatedWords

pub fn add(&self, other: CreatedWords) -> CreatedWords

pub fn has_word<P: Into<i64> + Copy>(&self, length: P) -> HasWord

pub fn is_empty(&self) -> bool

pub fn not_empty(&self) -> bool