sudachi::analysis::mlist

Struct MorphemeList

pub struct MorphemeList<T> { /* private fields */ }

Implementations

impl<T: DictionaryAccess> MorphemeList<T>

pub fn empty(dict: T) -> Self

Returns an empty morpheme list

pub fn from_components( dict: T, input: InputBuffer, path: Vec<ResultNode>, subset: InfoSubset, ) -> Self

Creates MorphemeList from components

pub fn collect_results<U: DictionaryAccess>( &mut self, analyzer: &mut StatefulTokenizer<U>, ) -> SudachiResult<()>

pub fn split_into( &self, mode: Mode, index: usize, out: &mut Self, ) -> SudachiResult<bool>

Splits morphemes and writes them into the resulting list The resulting list is not cleared before that Returns true if split produced more than two elements

pub fn clear(&mut self)

Clears morphemes from analysis result

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn get(&self, idx: usize) -> Morpheme<'_, T>

pub fn surface(&self) -> Ref<'_, str>

pub fn iter(&self) -> MorphemeIter<'_, T>

pub fn get_internal_cost(&self) -> i32

Gets the whole cost of the path

pub fn dict(&self) -> &T

pub fn subset(&self) -> InfoSubset

pub fn copy_slice(&self, start: usize, end: usize, out: &mut Self)

pub fn lookup( &mut self, query: &str, subset: InfoSubset, ) -> SudachiResult<usize>

impl<T: DictionaryAccess + Clone> MorphemeList<T>

pub fn empty_clone(&self) -> Self

pub fn split(&self, mode: Mode, index: usize) -> SudachiResult<MorphemeList<T>>

👎Deprecated since 0.6.1:
use split_into

Returns a new morpheme list splitting the morpheme with a given mode. Returns an empty list if there was no splits