Struct sudachi::analysis::morpheme::MorphemeList

pub struct MorphemeList<T> { /* fields omitted */ }

A list of morphemes

Implementations

impl<T: DictionaryAccess> MorphemeList<T>

pub fn from_components(
dict: T,
original: String,
path: Vec<ResultNode>
) -> SudachiResult<Self>

pub fn empty(dict: T) -> Self

Returns an empty morpheme list.

pub fn collect_results<U: DictionaryAccess>(
&mut self,
analyzer: &mut StatefulTokenizer<U>
) -> SudachiResult<()>

pub fn get_grammar(&self) -> &Grammar<'_>

impl<T: DictionaryAccess + Clone> MorphemeList<T>

pub fn split(&self, mode: Mode, index: usize) -> SudachiResult<MorphemeList<T>>

Returns a new morpheme list splitting the morpheme with a given mode.

impl<T> MorphemeList<T>

pub fn iter(&self) -> MorphemeIter<'_, T>

pub fn surface(&self) -> &str

pub fn get_node(&self, index: usize) -> &ResultNode

pub fn get_begin(&self, index: usize) -> usize

pub fn get_end(&self, index: usize) -> usize

pub fn get_surface(&self, index: usize) -> &str

Returns a substring of the original text which corresponds to the morpheme

pub fn get_word_info(&self, index: usize) -> &WordInfo

pub fn is_oov(&self, index: usize) -> bool

pub fn get_internal_cost(&self) -> i32

Returns the total cost of the path

pub fn len(&self) -> usize

pub fn get(&self, index: usize) -> Morpheme<'_, T>