Struct sudachi::dic::build::DictBuilder

pub struct DictBuilder<D> { /* fields omitted */ }

Builds a binary dictionary from csv lexicon and connection matrix (optional)

Implementations

impl DictBuilder<NoDic>

pub fn new_system() -> Self

Creates a new builder for system dictionary

impl<D: DictionaryAccess> DictBuilder<D>

pub fn new_user(system: D) -> Self

Creates a new builder for user dictionary

pub fn set_compile_time<T: Into<SystemTime>>(&mut self, time: T) -> SystemTime

Set the dictionary compile time to the specified time instead of current time

pub fn set_description<T: Into<String>>(&mut self, description: T)

Set the dictionary description

pub fn read_lexicon<'a, T: AsDataSource<'a> + 'a>(
&mut self,
data: T
) -> SudachiResult<usize>

Read the csv lexicon from either a file or an in-memory buffer

pub fn read_conn<'a, T: AsDataSource<'a> + 'a>(
&mut self,
data: T
) -> SudachiResult<()>

Read the connection matrix from either a file or an in-memory buffer

pub fn compile<W: Write>(&mut self, w: &mut W) -> SudachiResult<()>

Compile the binary dictionary and write it to the specified sink

pub fn resolve(&mut self) -> SudachiResult<usize>

Resolve the dictionary references.

Returns the number of resolved entries

pub fn report(&self) -> &[DictPartReport]

Return dictionary build report