Struct sudachi::dic::category_type::CategoryType

#[repr(transparent)]
pub struct CategoryType(_);

A set of categories for a character

Implemented as a bitset with fixed size

Implementations

impl CategoryType

pub const DEFAULT: Self = _

The fall back category.

pub const SPACE: Self = _

White spaces.

pub const KANJI: Self = _

CJKV ideographic characters.

pub const SYMBOL: Self = _

Symbols.

pub const NUMERIC: Self = _

Numerical characters.

pub const ALPHA: Self = _

Latin alphabets.

pub const HIRAGANA: Self = _

Hiragana characters.

pub const KATAKANA: Self = _

Katakana characters.

pub const KANJINUMERIC: Self = _

Kanji numeric characters.

pub const GREEK: Self = _

Greek alphabets.

pub const CYRILLIC: Self = _

Cyrillic alphabets.

pub const USER1: Self = _

User defined category.

pub const USER2: Self = _

User defined category.

pub const USER3: Self = _

User defined category.

pub const USER4: Self = _

User defined category.

pub const NOOOVBOW: Self = _

This character cannot be the beginning of an OOV word

pub const NOOOVBOW2: Self = _

This and next characters cannot be the beginning of an OOV word

pub const ALL: Self = _

All categories at once except NOOOVBOW/2

impl CategoryType

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const fn from_bits_retain(bits: u32) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub fn from_name(name: &str) -> Option<Self>

Get the value for a flag from its stringified name.

Names are case-sensitive, so must correspond exactly to the identifier given to the flag.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

This method is equivalent to union.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

This method is equivalent to difference.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

This method is equivalent to symmetric_difference.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Calculating self bitwise and (&) other, including any bits that don’t correspond to a defined flag.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Calculates self bitwise or (|) other, including any bits that don’t correspond to a defined flag.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Calculates self bitwise and (&!) the bitwise negation of other, including any bits that don’t correspond to a defined flag.

This method is not equivalent to a & !b when there are bits set that don’t correspond to a defined flag. The ! operator will unset any bits that don’t correspond to a flag, so they’ll always be unset by a &! b, but respected by a.difference(b).

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Calculates self bitwise exclusive or (^) other, including any bits that don’t correspond to a defined flag.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Calculates the bitwise negation (!) of self, unsetting any bits that don’t correspond to a defined flag.

impl CategoryType

pub const fn iter(&self) -> Iter<CategoryType>

Iterate over enabled flag values.

pub const fn iter_names(&self) -> IterNames<CategoryType>

Iterate over enabled flag values with their stringified names.

impl CategoryType

pub fn count(self) -> u32