Module sudachi::dic::lexicon

Modules

trie

word_id_table

word_infos

word_params

Structs

Lexicon

Dictionary lexicon

LexiconEntry

Result of the Lexicon lookup

Constants

MAX_DICTIONARIES

The first 4 bits of word_id are used to indicate that from which lexicon the word comes, thus we can only hold 15 lexicons in the same time. 16th is reserved for marking OOVs.