sudachi::dic

Module word_info

Re-exports

pub use parse::WordInfoParser;

Modules

layout

parse

Structs

WordInfo

WordInfo API.

WordInfoData

wrapper type that indicates inner data are resolved for the specific lexicon set.

WordInfoFixedData

WordInfoRawData

Parsed raw binary representation of a word info entry.

WordInfoRefData

wrapper type that indicates inner data are not resolved for the specific lexicon set.

WordInfos

Enums

WordInfoError

Errors returned while scanning binary WordInfo entry boundaries.

Traits

WordInfoResolver

Data structure needed to resolve references in the WordInfoData. Currently only lexicon set is needed, but it can be extended in the future if needed.