sudachi::dic::word_info

Struct WordInfos

pub struct WordInfos<'a> { /* private fields */ }

Implementations

impl<'a> WordInfos<'a>

pub const ENTRIES_INITIAL_OFFSET: usize = layout::ENTRY_INITIAL_OFFSET

pub const WORD_ID_ALIGNMENT_BITS: usize = layout::WORD_ID_ALIGNMENT_BITS

pub const WORD_INFO_OFFSET_ALIGNMENT: usize = layout::WORD_INFO_OFFSET_ALIGNMENT

pub fn from_bytes(bytes: &'a [u8]) -> WordInfos<'a>

pub fn entry_id_to_offset(entry_id: EntryId) -> usize

pub fn entry_ids_in_order(&self, num_total_entries: u32) -> Option<Vec<EntryId>>

pub fn validate_entry_boundaries( &self, num_total_entries: u32, ) -> SudachiResult<()>

Validate that all WordInfo entries can be scanned from their binary boundaries.

Returns WordInfoError through SudachiError when an entry id cannot be derived or an entry size cannot be read.

pub fn get_word_info( &self, entry_id: EntryId, subset: InfoSubset, ) -> SudachiResult<WordInfoRefData>