Module sudachi::analysis::node

Structs

NodeSplitIterator

ResultNode

Full lattice node, as the result of analysis. All indices (including inner) are in the original sentence space.

Traits

LatticeNode

PathCost

Accessor trait for the full path cost

RightId

Accessor trait for right connection id

Functions

concat_nodes

Concatenate the nodes in the range and replace normalized_form if given.

concat_oov_nodes

Concatenate the nodes in the range and set pos_id.