Struct sudachi::analysis::node::ResultNode

pub struct ResultNode { /* private fields */ }

Full lattice node, as the result of analysis. All indices (including inner) are in the modified sentence space Indices are converted to original sentence space when user request them.

Implementations

impl ResultNode

pub fn new( inner: Node, total_cost: i32, begin_bytes: u16, end_bytes: u16, word_info: WordInfo ) -> ResultNode

impl ResultNode

pub fn word_info(&self) -> &WordInfo

pub fn begin_bytes(&self) -> usize

Returns begin offset in bytes of node surface in a sentence

pub fn end_bytes(&self) -> usize

Returns end offset in bytes of node surface in a sentence

pub fn bytes_range(&self) -> Range<usize>

Returns range in bytes (for easy string slicing)

pub fn set_bytes_range(&mut self, begin: u16, end: u16)

pub fn set_char_range(&mut self, begin: u16, end: u16)

pub fn num_splits(&self, mode: Mode) -> usize

Returns number of splits in a specified mode

pub fn split<'a>( &'a self, mode: Mode, lexicon: &'a LexiconSet<'a>, subset: InfoSubset, text: &'a InputBuffer ) -> NodeSplitIterator<'a>

Split the node with a specified mode using the dictionary data