Struct sudachi::config::Config

pub struct Config {
    pub system_dict: Option<PathBuf>,
    pub user_dicts: Vec<PathBuf>,
    pub character_definition_file: PathBuf,
    pub connection_cost_plugins: Vec<Value>,
    pub input_text_plugins: Vec<Value>,
    pub oov_provider_plugins: Vec<Value>,
    pub path_rewrite_plugins: Vec<Value>,
    /* private fields */
}

Setting data loaded from config file

Fields
system_dict: Option<PathBuf>user_dicts: Vec<PathBuf>character_definition_file: PathBufconnection_cost_plugins: Vec<Value>input_text_plugins: Vec<Value>oov_provider_plugins: Vec<Value>path_rewrite_plugins: Vec<Value>
Implementations

impl Config

pub fn new(
config_file: Option<PathBuf>,
resource_dir: Option<PathBuf>,
dictionary_path: Option<PathBuf>
) -> Result<Self, ConfigError>

pub fn minimal_at(resource_dir: impl Into<PathBuf>) -> Config

Creates a minimal config with the provided resource directory

pub fn with_system_dic(self, system: impl Into<PathBuf>) -> Config

Sets the system dictionary to the provided path

pub fn resolve_paths(&self, path: String) -> Vec<String>

pub fn complete_path<P: AsRef<Path> + Into<PathBuf>>(
&self,
file_path: P
) -> Result<PathBuf, ConfigError>

Resolves a possibly relative path with regards to all possible anchors:

Absolute paths stay as they are

Paths are resolved wrt to anchors, returning the first existing one

Path are checked wrt to CWD

If all fail, return an error with all candidate paths listed

pub fn resolved_system_dict(&self) -> Result<PathBuf, ConfigError>

pub fn resolved_user_dicts(&self) -> Result<Vec<PathBuf>, ConfigError>