Struct sudachi::dic::dictionary::JapaneseDictionary

pub struct JapaneseDictionary { /* private fields */ }

Implementations

impl JapaneseDictionary

pub fn from_cfg(cfg: &Config) -> SudachiResult<JapaneseDictionary>

Creates a dictionary from the specified configuration Dictionaries will be read from disk

pub fn from_cfg_storage(
cfg: &Config,
storage: SudachiDicData
) -> SudachiResult<JapaneseDictionary>

Creats a dictionary from the specified configuration and storage

pub fn grammar<'a>(&'a self) -> &'_ Grammar<'a>

Returns grammar with the correct lifetime

pub fn lexicon<'a>(&'a self) -> &'_ LexiconSet<'a>

Returns lexicon with the correct lifetime