sudachi::dic::lexicon_set

Struct LexiconSet

pub struct LexiconSet<'a> { /* private fields */ }

Set of Lexicons

Handles multiple lexicons as one lexicon The first lexicon in the list must be from system dictionary

Implementations

impl<'a> LexiconSet<'a>

pub fn new(system_lexicon: Lexicon<'_>, num_system_pos: usize) -> LexiconSet<'_>

Creates a LexiconSet given a lexicon

It is assumed that the passed lexicon is the system dictionary

pub fn append( &mut self, lexicon: Lexicon<'a>, pos_offset: usize, ) -> Result<(), LexiconSetError>

Add a lexicon to the lexicon list

pos_offset: number of pos in the grammar

pub fn is_full(&self) -> bool

Returns if dictionary capacity is full

impl LexiconSet<'_>

pub fn lookup<'b>( &'b self, input: &'b [u8], offset: usize, ) -> impl Iterator<Item = LexiconEntry> + 'b

Returns iterator which yields all words in the dictionary, starting from the offset bytes

Searches dictionaries in the reverse order: user dictionaries first and then system dictionary

pub fn get_word_info(&self, id: WordId) -> SudachiResult<WordInfo>

Returns WordInfo for given WordId

pub fn get_word_info_subset( &self, id: WordId, subset: InfoSubset, ) -> SudachiResult<WordInfo>

Returns WordInfo for given WordId. Only fills a requested subset of fields. Rest will be of default values (0 or empty).

pub fn get_word_param(&self, id: WordId) -> (i16, i16, i16)

Returns word_param for given word_id

pub fn size(&self) -> u32